Get new version of Feed class into feedList branch for further work (#6)
* Implemented Item class to save the items of the feed and wrote saving algorithm.

I found the previous mistake, but it is still present. The problem is, after the get() function, before handing the reply over to parse(), we jump back into main and then call our tests. But logically, there is nothing to test on :D I'm not sure, if eventually a Timer in get() would fix this issue. I think it's because we go to parse() if the download is finished, and in the meanwhile we try to work further through main which isn't possible..

* Fixed the Timer issue, now it stays in get() until the download is finished.

The download works like we want it to now, but somehow it seems he saves every child of "item" in every variable like "link" "title" and so on..

* Fixed latest issue with saving the wrong data

* Fixed the html issue in "description". Merry Christmas :)

Co-authored-by: tmoSolutions <94002135+tmoSolutions@users.noreply.github.com>